requirements.txt
streamlit
fastapi
uvicorn
openai
langchain-openai
pydantic
python-dotenv